ISMPrakt04edit

+++ ISMPrakt04ClassLib/ForConsole1.cs
@@ -8,9 +8,9 @@ namespace ISMPrakt04ClassLib
 {
     public class ForConsole1
     {
-        public int Solve(int a=0, int b=0)
+        public double Solve(int a=0, int b=0)
         {
-            int res = 1;
+            double res = 1;
             for (; a <= b; a++) res *= a;
             return res;
         }
